requirements.txt
crewai==0.11.2
crewai-tools==0.0.1
langchain<0.2.0
langchain-community
ollama
python-dotenv==1.0.0
pydantic>=2.0.0
